@@ -12,15 +12,11 @@ android {
             storeFile file('C:\\Users\\Andy\\AndroidStudioProjects\\FastBle\\app\\build\\outputs\\apk\\release\\app-release-unsigned.apk')
         }
     }
-    compileSdkVersion 29
-    buildToolsVersion '27.0.3'
-    lintOptions{
-        disable 'MissingTranslation'
-    }
+    compileSdkVersion 34
     defaultConfig {
         applicationId "com.clj.blesamplePCBA"
-        minSdkVersion 14
-        targetSdkVersion 29
+        minSdkVersion 19
+        targetSdkVersion 34
         versionCode 235
         versionName "2.3.3"
         vectorDrawables.useSupportLibrary = true
@@ -34,6 +30,10 @@ android {
             debuggable true
         }
     }
+    namespace 'com.clj.blesamplePCBA'
+    lint {
+        disable 'MissingTranslation'
+    }
 }
 
 repositories {
@@ -46,11 +46,12 @@ dependencies {
     implementation 'com.android.support:design:27.1.1'
     //noinspection GradleCompatible
     implementation 'com.android.support:support-v4:27.1.1'
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    testCompile 'junit:junit:4.12'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'com.google.android.material:material:1.12.0'
+    testImplementation 'junit:junit:4.12'
     //noinspection GradleCompatible
-    compile 'com.android.support:appcompat-v7:27.1.1'
-    compile project(':FastBleLib')
+    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation project(':FastBleLib')
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
 
     //CHARTING
